Move existing files into System.Runtime.InteropServices.Analyzers
CA1401 and CA2101 analyzers, From PInvokeDiagnosticAnalyzer in System.Runtime.InteropServices.Analyzers
CA2101 fixer, from SpecifyMarshalingForPInvokeStringArguments.Fixer in System.Runtime.InteropServices.Analyzers
RS0015 analyzer, from ConsumePreserveSigAnalyzer in Roslyn.Diagnostics.Analyzers (no tests)
