+++ PortFxCop/src/System.Runtime.InteropServices.Analyzers/Core/AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttribute.cs
@@ -1,38 +1,81 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Collections.Immutable;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.Diagnostics;
 
-namespace System.Runtime.InteropServices.Analyzers
-{                   
-    /// <summary>
-    /// RS0015: Always consume the value returned by methods marked with PreserveSigAttribute
-    /// </summary>
-    public abstract class AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttributeAnalyzer : DiagnosticAnalyzer
+namespace Roslyn.Diagnostics.Analyzers
+{
+    public abstract class ConsumePreserveSigAnalyzer<TSyntaxKind> : DiagnosticAnalyzer
+        where TSyntaxKind : struct
     {
-        internal const string RuleId = "RS0015";
+        private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(RoslynDiagnosticsResources.ConsumePreserveSigTitle), RoslynDiagnosticsResources.ResourceManager, typeof(RoslynDiagnosticsResources));
+        private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(RoslynDiagnosticsResources.ConsumePreserveSigMessage), RoslynDiagnosticsResources.ResourceManager, typeof(RoslynDiagnosticsResources));
+        private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(RoslynDiagnosticsResources.ConsumePreserveSigDescription), RoslynDiagnosticsResources.ResourceManager, typeof(RoslynDiagnosticsResources));
 
-        private static readonly LocalizableString s_localizableTitle = new LocalizableResourceString(nameof(SystemRuntimeInteropServicesAnalyzersResources.AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttributeTitle), SystemRuntimeInteropServicesAnalyzersResources.ResourceManager, typeof(SystemRuntimeInteropServicesAnalyzersResources));
-        
-        private static readonly LocalizableString s_localizableMessage = new LocalizableResourceString(nameof(SystemRuntimeInteropServicesAnalyzersResources.AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttributeMessage), SystemRuntimeInteropServicesAnalyzersResources.ResourceManager, typeof(SystemRuntimeInteropServicesAnalyzersResources));
-        private static readonly LocalizableString s_localizableDescription = new LocalizableResourceString(nameof(SystemRuntimeInteropServicesAnalyzersResources.AlwaysConsumeTheValueReturnedByMethodsMarkedWithPreserveSigAttributeDescription), SystemRuntimeInteropServicesAnalyzersResources.ResourceManager, typeof(SystemRuntimeInteropServicesAnalyzersResources));
-        
-        internal static DiagnosticDescriptor Rule = new DiagnosticDescriptor(RuleId,
-                                                                             s_localizableTitle,
-                                                                             s_localizableMessage,
-                                                                             DiagnosticCategory.Reliability,
-                                                                             DiagnosticSeverity.Warning,
-                                                                             isEnabledByDefault: false,
-                                                                             description: s_localizableDescription,
-                                                                             helpLinkUri: null,     // TODO: add MSDN url
-                                                                             customTags: WellKnownDiagnosticTags.Telemetry);
+        internal static readonly DiagnosticDescriptor ConsumePreserveSigAnalyzerDescriptor = new DiagnosticDescriptor(
+            RoslynDiagnosticIds.ConsumePreserveSigRuleId,
+            s_localizableTitle,
+            s_localizableMessage,
+            "Reliability",
+            DiagnosticSeverity.Warning,
+            isEnabledByDefault: true,
+            description: s_localizableDescription,
+            customTags: WellKnownDiagnosticTags.Telemetry);
 
-        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Rule);
+        private INamedTypeSymbol _lazyPreserveSigType;
 
-        public override void Initialize(AnalysisContext analysisContext)
-        { 
-            
+        public sealed override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics
+        {
+            get
+            {
+                return ImmutableArray.Create(ConsumePreserveSigAnalyzerDescriptor);
+            }
+        }
+
+        public sealed override void Initialize(AnalysisContext context)
+        {
+            context.RegisterCompilationStartAction(compilationContext =>
+            {
+                _lazyPreserveSigType = compilationContext.Compilation.GetTypeByMetadataName("System.Runtime.InteropServices.PreserveSigAttribute");
+                if (_lazyPreserveSigType != null)
+                {
+                    compilationContext.RegisterSyntaxNodeAction(AnalyzeNode, ImmutableArray.Create(InvocationExpressionSyntaxKind));
+                }
+            });
+        }
+
+        protected abstract TSyntaxKind InvocationExpressionSyntaxKind { get; }
+        protected abstract bool IsExpressionStatementSyntaxKind(int rawKind);
+
+        private void AnalyzeNode(SyntaxNodeAnalysisContext context)
+        {
+            if (_lazyPreserveSigType == null)
+            {
+                return;
+            }
+
+            var node = context.Node;
+            if (!IsExpressionStatementSyntaxKind(node.Parent.RawKind))
+            {
+                return;
+            }
+
+            var symbol = context.SemanticModel.GetSymbolInfo(node, context.CancellationToken).Symbol;
+            if (symbol == null)
+            {
+                return;
+            }
+
+            foreach (var attributeData in symbol.GetAttributes())
+            {
+                if (attributeData.AttributeClass.Equals(_lazyPreserveSigType))
+                {
+                    var diagnostic = Diagnostic.Create(ConsumePreserveSigAnalyzerDescriptor, node.GetLocation(), symbol);
+                    context.ReportDiagnostic(diagnostic);
+                    return;
+                }
+            }
         }
     }
 }
