+++ PortFxCop/src/System.Runtime.InteropServices.Analyzers/Core/DocumentChangeAction.cs
