+++ PortFxCop/src/System.Runtime.InteropServices.Analyzers/Core/PInvokeDiagnosticAnalyzer.cs
