+++ PortFxCop/src/System.Runtime.InteropServices.Analyzers/Core/System.Runtime.InteropServices.Analyzers.props
@@ -1,15 +1,16 @@
 <Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 
-   <!-- 
+  <!-- 
     This file contains the rules that have been implemented in this package and therefore should be disabled for the binary FxCop.
     When a new rule is implemented in this package, make sure to add an entry below to disable that rule in FxCop to avoid duplicate 
     issues.
-    The format is -[Category]#[ID], e.g., -Microsoft.Design#CA1001;
+    The format is -[Category]#[ID]
   -->
-   <PropertyGroup>
-      <CodeAnalysisRuleSetOverrides>
-         $(CodeAnalysisRuleSetOverrides);
-         <!--INSERTRULESETOVERRIDES-->
-      </CodeAnalysisRuleSetOverrides>
-   </PropertyGroup>
+  <PropertyGroup>
+    <CodeAnalysisRuleSetOverrides>$(CodeAnalysisRuleSetOverrides);
+      -Microsoft.Design#CA1060;
+      -Microsoft.Interoperability#CA1401;
+      -Microsoft.Globalization#CA2101;
+    </CodeAnalysisRuleSetOverrides>
+  </PropertyGroup>
 </Project>
